fix: Change naming, fix NullArgumentException
